# server/docs/swagger.yaml
basePath: /api
definitions:
...
  cows.FilterSerializedCow:
    properties:
      approved:
        description: Подтверждена ли админом
        example: true
        type: boolean
      birkingDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата перебирковки
      birthDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата рождения коровы
      breedName:
        description: Название породы коровы
        example: Какая-нибудь порода
        type: string
      calvingDate:
        description: Дата отела
        items:
          $ref: '#/definitions/models.DateOnly'
        type: array
      checkMilkDate:
        description: Дата контрольного доения
        items:
          $ref: '#/definitions/models.DateOnly'
        type: array
      createdAt:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата внесения информации о корове в БД
      deathDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата смерти
      departDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата выбытия
      events:
        description: Вет события
        items:
          $ref: '#/definitions/models.Event'
        type: array
      exteriorRating:
        description: Оценка экстерьера
        type: number
      farmGroupName:
        description: Название хозяйства, в котором корова
        example: ООО Аурус
        type: string
      genotyped:
        description: Факт генотипирования коровы
        example: true
        type: boolean
      genotypingDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата генотипирования
      hozName:
        description: Название хозяйства
        type: string
      id:
        description: ID коровы
        example: 123
        type: integer
      inbrindingCoeffByFamily:
        description: Коэф. инбриндинга по родословной
        example: 3.14
        type: number
      inbrindingCoeffByGenotype:
        description: Коэф. инбриндинга по генотипу
        example: 3.14
        type: number
      insemenationDate:
        description: Дата осеменения
        items:
          $ref: '#/definitions/models.DateOnly'
        type: array
      inventoryNumber:
        description: Инвентарный номер коровы
        example: "321"
        type: string
      isAborted:
        description: Факт аборта
        type: boolean
      isDead:
        description: Факт смерти
        type: boolean
      isGenotyped:
        description: Факт генотипирования
        type: boolean
      isStillBorn:
        description: Факт мертворождения
        type: boolean
      isTwins:
        description: Факт рождения близнецов
        type: boolean
      monogeneticIllneses:
        description: моногенные заболевания
        items:
          $ref: '#/definitions/models.GeneticIllnessData'
        type: array
      name:
        description: Кличка коровы
        example: Буренка
        type: string
      rshnnumber:
        description: РСХН номер коровы
        example: "123"
        type: string
      sexName:
        description: Название породы
        type: string
    required:
    - approved
    - birthDate
    - farmGroupName
    - genotyped
    - id
    - inventoryNumber
    - name
    - rshnnumber
    type: object
...
  cows.ReserealizedCow:
    properties:
      approved:
        description: Целое число, 0 - корова не подтверждена, 1 - корова подтверждена,
          -1 - корова отклонена
        example: 1
        type: integer
      birkingDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата перебирковки
      birthDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: День рождения
      birthHozId:
        description: ID хозяйства рождения
        type: integer
      birthMethod:
        description: 'способ зачатия: клон, эмбрион, искусственное осеменени, естественное
          осеменение'
        type: string
      breedId:
        description: ID породы коровы
        example: 1
        type: integer
      breedName:
        description: порода, null если нет
        type: string
      createdAt:
        description: Время создания коровы в базе данных
        example: "2007-01-01"
        type: string
      deathDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата смерти
      departDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: День отбытия из коровника
      exterior:
        $ref: '#/definitions/models.Exterior'
      farmGroupId:
        description: ID хозяйства, которому корова принадлежит
        example: 1
        type: integer
      farmID:
        description: ID фермы, которой корова принадлежит
        example: 1
        type: integer
      farmName:
        description: ферма на которой живет, null если нет
        type: string
      father:
        allOf:
        - $ref: '#/definitions/models.Cow'
        description: Отец
      fatherSelecs:
        description: ID коровы отца коровы
        type: integer
      genetic:
        $ref: '#/definitions/models.Genetic'
      gradeHozId:
        description: оценка по хозяйству
        example: 1
        type: integer
      gradeRegionId:
        description: оценка по региону
        example: 1
        type: integer
      holding:
        $ref: '#/definitions/models.Farm'
      holdingID:
        description: ID холдинга, которому принадлежит корова
        type: integer
      hozHame:
        description: хозяйство на котором живет, null, если нет
        type: string
      id:
        description: ID коровы
        example: 1
        type: integer
      identificationNumber:
        description: он все-таки есть! это какой-то не российский номер коровы
        type: string
      inbrindingCoeffByFamily:
        description: Exterior                float64  `example:"3.14"` // Оценка экстерьера
          коровы, будет переделано в ID экстерьера коровы
        example: 3.14
        type: number
      inventoryNumber:
        description: Инвентарный номер коровы
        example: "1213321"
        type: string
      mother:
        allOf:
        - $ref: '#/definitions/models.Cow'
        description: Мать
      motherSelecs:
        description: ID коровы матери коровы
        type: integer
      name:
        description: Кличка коровы
        example: Дима
        type: string
      previousHozId:
        description: ID предыдущего хозяйства, когда корову продают, она переходит
          к новому владельцу и становится "новой коровой"
        type: integer
      rshnnumber:
        description: РСХН номер коровы
        example: "1323323232"
        type: string
      selecsNumber:
        description: Селекс номер коровы
        example: 98989
        type: integer
      sexId:
        description: ID пола коровы
        example: 1
        type: integer
      sexName:
        description: пол, null если нет
        type: string
    type: object
...
  models.DateOnly:
    properties:
      time.Time:
        type: string
    type: object
...
  models.Exterior:
    properties:
      acrumLength:
        type: number
      bodyDepth:
        type: number
      bodyStructure:
        type: number
      boneQHockJointRear:
        type: number
      centerLigamentDepth:
        type: number
      chestWidth:
        type: number
      conditioning:
        type: number
      cowID:
        type: integer
      exteriorType:
        type: number
      foreLegPosFront:
        type: number
      foreTeatDiameter:
        type: number
      foreTeatLendth:
        type: number
      foreUdderAttach:
        type: number
      foreUdderPlcRear:
        type: number
      foreUdderWidth:
        type: number
      harmonyOfMovement:
        type: number
      heightOfUdderAttach:
        type: number
      hindLegPosRead:
        type: number
      hindLegPosSide:
        type: number
      hindTeatPlc:
        type: number
      hindUdderWidth:
        type: number
      hoofAngle:
        type: number
      id:
        type: integer
      limbs:
        type: number
      milkStrength:
        type: number
      pelvicWidth:
        type: number
      picturePath:
        type: string
      prominenceOfMilkVeins:
        type: number
      rating:
        type: number
      rearTeatDiameter:
        type: number
      rearTeatLength:
        type: number
      sacrumAngle:
        type: number
      sacrumHeight:
        type: number
      topLine:
        type: number
      udder:
        type: number
      udderBalance:
        type: number
      udderDepth:
        type: number
    type: object
...
  models.Genetic:
    properties:
      bloodDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата взятия пробы крови
      cowID:
        description: ID коровы
        type: integer
      geneticIllnessesData:
        description: Список генетических заболеваний, пустой если нет
        items:
          $ref: '#/definitions/models.GeneticIllnessData'
        type: array
      gtcFilePath:
        description: путь к gtc файлу относительно genmilk.ru/api/static/gtc
        type: string
      id:
        description: ID записи о генотипировании
        type: integer
      inbrindingCoeffByGenotype:
        description: Коэф. инбриндинга по генотипу
        type: number
      probeNumber:
        description: Номер пробы
        type: string
      resultDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата получения  результата
    type: object
...
  models.Grade:
    properties:
      ebvFat:
        description: Оценка жира по EBV
        type: number
      ebvInsemenation:
        description: Оценка кратности осеменения по EBV
        type: number
      ebvMilk:
        description: Оценка удоя по EBV
        type: number
      ebvProtein:
        description: Оценка белка по EBV
        type: number
      evbService:
        description: Оценка длительности сервисного периода по EBV
        type: number
      generalValue:
        description: CowID           uint
        type: number
      id:
        description: ID оценки
        example: 1
        type: integer
    type: object
...
  models.Lactation:
    properties:
      abort:
        description: был ли аборт
        type: boolean
      calvingCount:
        description: 'Количество рожденных телят: 0 - мертворождение, 2 - двойня'
        type: integer
      calvingDate:
        $ref: '#/definitions/models.DateOnly'
      cowId:
        description: ID коровы, данные о лактации которой записаны
        type: integer
      days:
        description: количество дней, когда корова дает молоко
        type: integer
      fat305:
        description: Суммарный жир за 305 дней
        type: number
      fatAll:
        description: Суммарный жир
        type: number
      id:
        type: integer
      insemenationDate:
        allOf:
        - $ref: '#/definitions/models.DateOnly'
        description: Дата осеменения
      insemenationNum:
        description: Количество осеменений
        type: integer
      milk305:
        description: Суммарный надой за 305 дней
        type: number
      milkAll:
        description: Суммарный надой
        type: number
      number:
        description: номер лактации
        type: integer
      protein305:
        description: Суммарный белок за 305 дней
        type: number
      proteinAll:
        description: Суммарный белок
        type: number
      servicePeriod:
        description: 'сервис период коровы: время от отела до осеменения'
        type: integer
    type: object
...
host: genmilk.ru
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Сваггер сгенерирован из кода, поэтому может содержать неточности. По мере возможности они будут описаны далее
    Все даты передаются как строки
    Большая часть рутов не возвращает вложенные объекты
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GenMilk API
  version: "1.0"
paths:
...
  /cows:
    get:
      description: Возращает коров удовлетворяющих условиям фильтрации.
      parameters:
      - description: ID фермы (НЕ хозяйства), к которой принадлежит корова
        in: query
        name: farm_id
        type: integer
      - description: ID хозяйства (НЕ фермы), к которому принадлежит корова
        in: query
        name: farm_group_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cows.ReserealizedCow'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get list of cows
      tags:
      - Cows
  /cows/{id}:
    get:
      description: Возращает конкретную корову. Поля Father и Mother, имеют FatherId
        и MotherID null всегда, это неправильно, но так надо
      parameters:
      - description: ID конкретной коровы, чтобы ее вернуть
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cows.ReserealizedCow'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get concrete cow
      tags:
      - Cows
...
  /cows/{id}/genetic:
    get:
      description: Возращает генетическую информацию для коровы, null, если нет
      parameters:
      - description: ID коровы для которой ищется генетическая информация
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Genetic'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get list of check milks
      tags:
      - Cows
  /cows/{id}/grades:
    get:
      description: Возращает словарь с двумя ключам "ByRegion" - оценки по региону
        и "ByHoz" - оценки по хозяйству
      parameters:
      - description: ID коровы для которой ищутся оценки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/models.Grade'
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get grades
      tags:
      - Cows
...
  /cows/{id}/lactations:
    get:
      description: Возращает список всех лактаций для конкретной коровы.
      parameters:
      - description: ID коровы для которой ищутся лактации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Lactation'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get list of lactations
      tags:
      - Cows
  /cows/filter:
    post:
      consumes:
      - application/json
      description: |-
        Возращает словарь с двумя ключами "N", "LST". По ключу "N" - общее кол-во найденных коров,
        по ключу "LST" массив объектов filterSerealizedCow (см. Models)
      parameters:
      - description: applied filters
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/cows_filter.CowsFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties:
                $ref: '#/definitions/cows.FilterSerializedCow'
              type: object
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get filtered list of cows
      tags:
      - Cows
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"